+++ src/vs/workbench/contrib/notebook/common/notebookCommon.ts
@@ -36,14 +36,6 @@ export enum CellOutputKind {
 	Rich = 3
 }
 
-export enum KernelsLanguage {
-	SQL = 'sql',
-	Python = 'python',
-	PySpark = 'python',
-	SparkScala = 'scala',
-	SparkR = 'sparkr',
-	PowerShell = 'powershell'
-}
 
 export const NOTEBOOK_DISPLAY_ORDER = [
 	'application/json',
